% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_missing_weeks}
\alias{add_missing_weeks}
\title{A침adir las semanas epidemiol칩gicas faltantes}
\usage{
add_missing_weeks(dataset, col_epiweek)
}
\description{
A침adir las semanas epidemiol칩gicas faltantes
}
