% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_colnames_spaces}
\alias{clean_colnames_spaces}
\title{Función para reemplazar espacios en los nombres de las columnas con
guiones bajos}
\usage{
clean_colnames_spaces(df)
}
\description{
Función para reemplazar espacios en los nombres de las columnas con
guiones bajos
}
