% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_data_other_viruses}
\alias{clean_data_other_viruses}
\title{Limpiar los datos de la base de Otros Virus}
\usage{
clean_data_other_viruses(report_data)
}
\description{
Limpiar los datos de la base de Otros Virus
}
