% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_filmarray_data}
\alias{clean_filmarray_data}
\title{Limpiar los datos de la base de la Fundación Cardio Infantil}
\usage{
clean_filmarray_data(filmarray_data)
}
\description{
Limpiar los datos de la base de la Fundación Cardio Infantil
}
