% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_tosferina_data}
\alias{clean_tosferina_data}
\title{Limpiar los datos de la base de Tosferina}
\usage{
clean_tosferina_data(report_data)
}
\description{
Limpiar los datos de la base de Tosferina
}
