% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{complete_age_categories}
\alias{complete_age_categories}
\title{Completar las categorías de edad}
\usage{
complete_age_categories(data_grouped, event_name, event_label)
}
\description{
Completar las categorías de edad
}
