% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{generate_age_groups_viruses}
\alias{generate_age_groups_viruses}
\title{Generar los grupos de edad para la base de datos Otro Virus}
\usage{
generate_age_groups_viruses(
  report_data,
  event_name = "adenovirus",
  wt_percentage = FALSE,
  total_cases = 0,
  event_label
)
}
\description{
Generar los grupos de edad para la base de datos Otro Virus
}
