% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{get_all_tables}
\alias{get_all_tables}
\title{Obtener todas las tables de las bases historicas}
\usage{
get_all_tables(file_name, sheet_name)
}
\description{
Obtener todas las tables de las bases historicas
}
