% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_cases_epiweeks}
\alias{get_cases_epiweeks}
\title{Obtener la distribuci贸n de casos por semana epidemiol贸gica}
\usage{
get_cases_epiweeks(
  report_data,
  data_grouped,
  col_epiweek,
  diseases_epiweek,
  table = FALSE
)
}
\description{
Obtener la distribuci贸n de casos por semana epidemiol贸gica
}
