% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_cases_filmarray}
\alias{get_cases_filmarray}
\title{Obtener los casos de los virus de la base de datos filmarray de
base de datos de la Fundación Cardio Infantil}
\usage{
get_cases_filmarray(
  report_data,
  positive_value = "DETECTADO",
  age_groups = TRUE,
  epiweek = NULL,
  vrs_influenza = NULL
)
}
\description{
Obtener los casos de los virus de la base de datos filmarray de
base de datos de la Fundación Cardio Infantil
}
