% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_cases_other_viruses}
\alias{get_cases_other_viruses}
\title{Obtener la distribución de casos de la base de datos Otros Virus}
\usage{
get_cases_other_viruses(
  report_data,
  epiweek = NULL,
  age_groups = FALSE,
  vrs_influenza = NULL
)
}
\description{
Obtener la distribución de casos de la base de datos Otros Virus
}
