% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_cases_prop_epiweek}
\alias{get_cases_prop_epiweek}
\title{Obtener la proporci贸n acumulada de los virus  por semana
epidemiol贸gica}
\usage{
get_cases_prop_epiweek(
  filmarray_data,
  other_virs_data,
  include_sars = TRUE,
  epiweek
)
}
\description{
Obtener la proporci贸n acumulada de los virus  por semana
epidemiol贸gica
}
