% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_cases_sars}
\alias{get_cases_sars}
\title{Obtener la distribución de casos de Sars CoV 2}
\usage{
get_cases_sars(
  report_data,
  positive_value = "DETECTADO",
  age_groups = FALSE,
  epiweek = NULL
)
}
\description{
Obtener la distribución de casos de Sars CoV 2
}
