% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_cases_tosferina}
\alias{get_cases_tosferina}
\title{Obtener la distribución de casos de Tosferina}
\usage{
get_cases_tosferina(
  report_data,
  result = "positivo",
  column = "semana_epidemiologica",
  col_results = "interpretacion_del_resultado"
)
}
\description{
Obtener la distribución de casos de Tosferina
}
