% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_colors_age_groups}
\alias{get_colors_age_groups}
\title{Obtener los colores de los grupos de edad}
\usage{
get_colors_age_groups(order = FALSE, hex_cods = FALSE, include_sars = FALSE)
}
\description{
Obtener los colores de los grupos de edad
}
