% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cons_viruses_age_text}
\alias{get_cons_viruses_age_text}
\title{Obtener el texto del consolidado de los virus por grupo de edad}
\usage{
get_cons_viruses_age_text(dataset, col_name = "grupo_edad", text_group)
}
\description{
Obtener el texto del consolidado de los virus por grupo de edad
}
