% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_dist_fci_other_vrs}
\alias{get_dist_fci_other_vrs}
\title{Obtener la distribuci贸n de casos de las base de datos Filmarray
de la Fundaci贸n Cardio Infantil y Otros Virus}
\usage{
get_dist_fci_other_vrs(fci_data, vrs_data)
}
\description{
Obtener la distribuci贸n de casos de las base de datos Filmarray
de la Fundaci贸n Cardio Infantil y Otros Virus
}
