% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_distribution_age_vr_sars}
\alias{get_distribution_age_vr_sars}
\title{Obtener la distribución de casos de las bases de datos Filmarray y
Otros Virus}
\usage{
get_distribution_age_vr_sars(data_vr, data_sars)
}
\description{
Obtener la distribución de casos de las bases de datos Filmarray y
Otros Virus
}
