% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_distribution_epiweek}
\alias{get_distribution_epiweek}
\title{Obtener la distribuci贸n de casos por semana epidemiol贸gicas}
\usage{
get_distribution_epiweek(
  report_data,
  positive_value = "DETECTADO",
  col_epiweek = "se"
)
}
\description{
Obtener la distribuci贸n de casos por semana epidemiol贸gicas
}
