% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_distribution_esi_sars}
\alias{get_distribution_esi_sars}
\title{Obtener la distribución de casos ESI incluyendo Sars CoV 2
Otros Virus}
\usage{
get_distribution_esi_sars(report_data, epiweek = 0, test = "bio_molecular")
}
\description{
Obtener la distribución de casos ESI incluyendo Sars CoV 2
Otros Virus
}
