% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_distribution_surveillance}
\alias{get_distribution_surveillance}
\title{Obtener la distribución de casos de la vigilancia
Otros Virus}
\usage{
get_distribution_surveillance(
  report_data,
  epiweek = 0,
  include_sars = FALSE,
  surveillance_type = "esi",
  test = NULL
)
}
\description{
Obtener la distribución de casos de la vigilancia
Otros Virus
}
