% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_historic_epi_times}
\alias{get_historic_epi_times}
\title{Obtener los tiempos epidemiologicos de los datos historicos}
\usage{
get_historic_epi_times(tabla)
}
\description{
Obtener los tiempos epidemiologicos de los datos historicos
}
