% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_max_age_group}
\alias{get_max_age_group}
\title{Obtener el grupo de edad con mayor número de casos}
\usage{
get_max_age_group(dataset, col_name)
}
\description{
Obtener el grupo de edad con mayor número de casos
}
\keyword{internal}
