% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_perc_viruses_age}
\alias{get_perc_viruses_age}
\title{Obtener el porcentaje por grupo de edad}
\usage{
get_perc_viruses_age(dataset, col_name = "grupo_edad")
}
\description{
Obtener el porcentaje por grupo de edad
}
