% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_prop_text}
\alias{get_prop_text}
\title{Obtener el texto de la proporción acumulada con Sars CoV 2}
\usage{
get_prop_text(dataset)
}
\description{
Obtener el texto de la proporción acumulada con Sars CoV 2
}
