% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_results_tosferina}
\alias{get_results_tosferina}
\title{Obtener los resultados de casos de Tosferina}
\usage{
get_results_tosferina(
  report_data,
  results = "positivo",
  columns = "semana_epidemiologica",
  col_results = "interpretacion_del_resultado"
)
}
\description{
Obtener los resultados de casos de Tosferina
}
