% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{get_selected_table}
\alias{get_selected_table}
\title{Extraer una tabla específica de la lista y devolverla como data.frame}
\usage{
get_selected_table(tables, INDICADOR)
}
\description{
Extraer una tabla específica de la lista y devolverla como data.frame
}
