% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_subtypes_values}
\alias{get_subtypes_values}
\title{Obtener los valores de los subtipos de un virus}
\usage{
get_subtypes_values(viruses, subtypes)
}
\description{
Obtener los valores de los subtipos de un virus
}
