% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table_epiweek_tosferina}
\alias{get_table_epiweek_tosferina}
\title{Obtener la tabla de casos por semana epidemiológica de Tosferina}
\usage{
get_table_epiweek_tosferina(report_data, epiweek)
}
\description{
Obtener la tabla de casos por semana epidemiológica de Tosferina
}
