% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tosferina_text_sex}
\alias{get_tosferina_text_sex}
\title{Obtener el texto de tosferina por grupo de edad}
\usage{
get_tosferina_text_sex(dataset, figure)
}
\description{
Obtener el texto de tosferina por grupo de edad
}
