% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_viruses_cumulative_proportion}
\alias{get_viruses_cumulative_proportion}
\title{Obtener la proporción acumulada de los virus}
\usage{
get_viruses_cumulative_proportion(report_data)
}
\description{
Obtener la proporción acumulada de los virus
}
