% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_columns_total}
\alias{group_columns_total}
\title{Agrupar los datos por columnas, ordenarlas y definir un
porcentaje}
\usage{
group_columns_total(
  disease_data,
  event_name = "adenovirus",
  col_names,
  wt_percentage = FALSE,
  total_cases = 0,
  event_label = NULL,
  sum_cases = FALSE,
  col_order = NULL,
  etiqueta = TRUE
)
}
\description{
Agrupar los datos por columnas, ordenarlas y definir un
porcentaje
}
