% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labrep-package.R
\docType{package}
\name{labrep-package}
\alias{labrep}
\alias{labrep-package}
\title{labrep: Paquete para la Generación Automática del Informe de Circulación Viral}
\description{
Herramienta para la Generación Automática del Informe de Circulación Viral
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/TRACE-LAC/labrep/issues}
}

}
\author{
\strong{Maintainer}: Geraldine Gómez-Millán \email{geralidine.gomez@javeriana.edu.co} (\href{https://orcid.org/0009-0007-8701-0568}{ORCID}) [contributor]

}
\keyword{internal}
