% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_age_group_distribution}
\alias{plot_age_group_distribution}
\title{Graficar la distribución de casos por grupos de edad}
\usage{
plot_age_group_distribution(
  report_data,
  var_x = "grupo_edad",
  var_y = "porcentaje",
  var_fill = "etiqueta",
  stacked_percentage = TRUE,
  include_sars = FALSE
)
}
\description{
Graficar la distribución de casos por grupos de edad
}
