% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_cumulative_proportion}
\alias{plot_cumulative_proportion}
\title{Graficar la distribución acumulada de los casos}
\usage{
plot_cumulative_proportion(data_proportion)
}
\description{
Graficar la distribución acumulada de los casos
}
