% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_distribution_epiweek}
\alias{plot_distribution_epiweek}
\title{Graficar la distribuci贸n de casos por semana epidemiol贸gica}
\usage{
plot_distribution_epiweek(
  report_data,
  var_x = "semanaepidemiologicavegeneral",
  var_y = "casos",
  var_fill = "etiqueta",
  influenza = FALSE,
  positives = NULL
)
}
\description{
Graficar la distribuci贸n de casos por semana epidemiol贸gica
}
