% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_distribution_surveillance}
\alias{plot_distribution_surveillance}
\title{Graficar la distribución de casos de la vigilancia}
\usage{
plot_distribution_surveillance(
  report_data,
  var_x = "casos",
  var_y = "porcentaje",
  var_fill = "etiqueta"
)
}
\description{
Graficar la distribución de casos de la vigilancia
}
