% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_historic_epi_time}
\alias{plot_historic_epi_time}
\title{Graficar el tiempo epidemiológico historico}
\usage{
plot_historic_epi_time(stacked_data, tabla, line_data)
}
\description{
Graficar el tiempo epidemiológico historico
}
