% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_results_tosferina}
\alias{plot_results_tosferina}
\title{Graficar la distribución de casos de Tosferina}
\usage{
plot_results_tosferina(
  report_data,
  column = "semana_epidemiologica",
  label_x = "Semana epidemiologica",
  positives = NULL,
  show_values = FALSE
)
}
\description{
Graficar la distribución de casos de Tosferina
}
