% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_table_epiweek_tosferina}
\alias{plot_table_epiweek_tosferina}
\title{Graficar la tabla con la distribuci贸n de casos de Tosferna por semana
epidemiol贸gica}
\usage{
plot_table_epiweek_tosferina(
  data_epiweek,
  distribution_epiweek,
  col_epiweek = "SE",
  epiweek
)
}
\description{
Graficar la tabla con la distribuci贸n de casos de Tosferna por semana
epidemiol贸gica
}
