% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data.R
\name{plot_table_vrs_epiweek}
\alias{plot_table_vrs_epiweek}
\title{Graficar la tabla con la distribuci贸n de casos por semana
epidemiol贸gica}
\usage{
plot_table_vrs_epiweek(data_epiweek, col_epiweek = "Semana", epiweek)
}
\description{
Graficar la tabla con la distribuci贸n de casos por semana
epidemiol贸gica
}
